FIX repaint bug

+++ sources/item.cpp
@@ -136,6 +136,6 @@ void Item::moveBy(qreal dx, qreal dy)
 void Item::repaint()
 {
     QGraphicsPixmapItem::setPos(
-      static_cast<int>(m_scale * m_position.x()) + m_borderLeft,
-      static_cast<int>(m_scale * m_position.y()) + m_borderTop);
+        m_scale * m_position.x() + m_borderLeft,
+        m_scale * m_position.y() + m_borderTop);
 }
